Expanding the bootstrap
See that I have some gaps at the moment in the moment where I need to set up the relatioships between the objects better. Not sure if the eager loading is helpful right now and need to figure out appropriate uses for this. Previous experience was just with a propriatary JPA implementation