Adding the controller
Also was sidetracked as I wanted to create the controller with Kotlin. Used the latest version but the plugin for intellij is not supporting it yet. Moving the version down to the latest supported Kotlin version gave me issue with using java 21 as only up to 19 is supported. Decided to focus on moving the demo project forward and seeing if I can use Kotlin a bit later